requirements.txt
torch>=2.2.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.30.0
evaluate>=0.4.1
rouge_score>=0.1.2
nltk>=3.8.1
tqdm>=4.66.0
PyYAML>=6.0
